Publishing branch for react and node
Node server is in the root folder, just run npm start. To run the react client, open a new terminal and navigate to the tesla-client folder in the root folder and run npm start. then the react app should run on localhost:3000 and json request should be proxied to the express api server on localhost:3001