React and Redux
The login page with react and redux now works. I grabbed the new sass files from the dev branch and updated the styling. Login theoretically works now and is storing the received data in the global store. The header communicates with the global store to retrieve vehicle name and charge data.