added local state updating for sliders and buttons
the sliders and buttons now call temporary functions (such as alerts and stuff) and the visual states update client side for all buttons and sliders. All that we need to do to make these work is call the correct API function.
A few other things need to be implemented still such as log out and waking up the vehicle.
Made the server send an entire fake vehicle object that is fully populated.