Changes to climate modal
Changed input range to a custom slider react component. This means it will need to be restyled, but it was necessary due to the fact that react rendering methods rely on states. This custom component has a special onChangeComplete function that can be called while onChange can update the state locally for visual updates. This will prevent the API from being queried every time the slider is moved a tiny bit, and only query once it is complete. I also implemented ways to detect the temperature display settings (F or C) in the climate modal and automatically convert everything. Charge modal still needs the new component.