Added global timer and new charge range input
charge range input has been updated to allow an update after done inputting. The global timer runs every 10 seconds and can call a function in Timer.js that will get the current vehicle object and store it. Added a function to all of the controls that will reset the timer back to 10 after a button is pressed. This is because if someone were to press a button when the timer had 1 second left, and the API got sent out, the client would update visually. Because the calls are asynchronous though, the server could receive the getvehicle information before the button call. This would result in the client going back to a previous state before the button was called, and then go out of sync with where the server is causing confusion for the user, thinking their call didnt go through when infact it did.