Mobile access denied and logout
Added a mobile access check that goes through and checks with each refresh, so if the vehicle turns it off the app will show that it's turned off, and if it's turned on, the app will allow access. Also added a logout button. We should restyle the logout button and place it inside the header bar.