actually fixed sunroof
now it checks to see if theres a sunroof and if there is, the button will appear, if not the button stays hidden. The button also reacts to if the sunroof is actually open now, so if the sunroof is present and open, then the button is labeled close, and it sends the close api call and vice versa. Can't physically test it, but it works sending the information to enable it in the fake vehicle.