Changing car images
images now load based on the vehicle. now it will load the appropriate image for the vehicle that is being logged into. Also edited the heartbeat routine for updating the vehicle information when the car is asleep so it shouldn't request another wake up again until the vehicle is either woken up or the request fails.