Breaking up main page and modal components
This allows us to individually control when any modal is shown based on the global state. previously control modals and stuff had been tied to their own component. This change allows me to close the modal behind the "are you sure" confirmation prompts to clean up the look and then bring them back upon clicking x on the confirmation modal or pressing yes. It keeps the modals from layering up on eachother.

Also prevented buttons from loading before the car loaded. Header now states the vehicle is loading when it is loading, and waking up the vehicle if it needs to wake it up.