Changed login procedure
Now when the server logs in, it doesn't treat all inputs as test, and it serves an invalid login prompt when the login isn't successful. To log into a test vehicle, the email must be "test", the password can be whatever, and then we can log into the test vehicle. The login prompt now also will put up a message saying login failed when the login fails, and it will automatically empty the email and password field for you to try again.