Fixed slider input issue on first click
Sliders were actually being set improperly on the first click down but worked once they were moved, this resulted in weird visual issues when the slider was clicked but not actually moved yet. After a move it would resolve itself. This was fixed. The problem was due to the asynchronous nature of setting the local state of a component versus the synchronous setting of the global state.