More graceful api call fails
Now the error messages for not retrieving vehicle data, or not getting the mobile access state are on a counter. They count individually. If it fails 5 or more times in a row, the error pops up. If it gets a successful response, the counter is reset to 0. This means that missing an API refresh call or two won't pop the error message up, and instead just rely on future pulls to resolve the issue silently.